fix(tuttid): split issue run lifecycle out of issues.go and un-rot the codex title scan fixture
issues.go crossed the 800-line revive limit after the orchestrator work;
the run lifecycle and planning-conversation notifications move to
issues_runs.go unchanged. The codex SQLite title scan test hardcoded
2026-06-18 timestamps while the scan applies a rolling 30-day cutoff, so
the fixture aged out of the window exactly 30 days later and the scan
silently returned nothing; the fixture now stamps relative to now.

(cherry picked from commit fc00d49d37fa991288a985109e9cb51d62b36e6c)
